Problem Decomposition using Multi-layered Simplex-lattice Design

Problem Decomposition using Multi-layered Simplex-lattice Design for MOEADr package

decomposition_msld(decomp, ...)

Arguments

decomp
list containing the relevant decomposition parameters. Besides decomp$name = "msld", this method requires the definition of the following key-value pairs in decomp:

decomp$H: array of positive integers representing the H values to be used by the SLD decomposition at each layer (see decomposition_sld() for details).

decomp$tau: array of scale multipliers for each layer, \(0 < \tau_i \le 1\), \(\tau_i != \tau_j\) for all \(i != j\). Must have the same length as decomp$H.

decomp$.nobj: integer value, decomp$.nobj > 1. Number of objectives of the problem.

...
other parameters (included for compatibility with generic call)

Details

This routine calculates the weight vectors for the MOEA/D using the Multi-layered Simplex-lattice Design.

References

K. Li et al. (2014), "An Evolutionary Many-Objective Optimization Algorithm Based on Dominance and Decomposition", IEEE Trans. Evol. Comp. 19(5):694-716, 2015. DOI: 10.1109/TEVC.2014.2373386

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi: 10.18637/jss.v092.i06

Examples

decomp <- list(name = "msld", H = c(5, 3), tau = c(.9, .5), .nobj = 4)
W <- decomposition_msld(decomp)