Problem Decomposition using Simplex-lattice Design

Problem Decomposition using Simplex-lattice Design for MOEADr package

decomposition_sld(decomp, ...)

Arguments

decomp
list containing the relevant decomposition parameters. Besides decomp$name = "sld", this method requires the definition of the following key-value pairs:

decomp$H, decomposition constant. Suggested values for decomp$H are (use with caution):

m	|	H	|	N

2	|	99	|	100

3	|	12	|	91

5	|	6	|	210

It is important to highlight that the number of vectors generated (N) must be greater than the number of neighbors declared in neighbors$T (see moead() for details).

decomp$.nobj: integer value, decomp$.nobj > 1. Number of objectives of the problem.

...
other parameters (included for compatibility with generic call)

Details

This routine calculates the weight vectors for the MOEA/D using the Simplex-lattice Design.

References

I. Das, J. Dennis (1998), "Normal Boundary Intersection - A New Method for Generating the Pareto Surface in Nonlinear Multicriteria Optimization Problems", SIAM J. Optim., 8(3), 631-657. DOI: 10.1137/S1052623496307510

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi: 10.18637/jss.v092.i06

Examples

decomp <- list(name = "sld", H = 99, .nobj = 2)
W <- decomposition_sld(decomp)