Problem Decomposition using Uniform Design

Problem Decomposition using Uniform Design for MOEADr package

decomposition_uniform(decomp, ...)

Arguments

decomp

list containing the relevant decomposition parameters. Besides decomp$name = "uniform", this method requires the definition of the following key-value pairs:

decomp$N, number of subproblems to generate. It is important to highlight that the number of subproblems must be greater than the number of neighbors declared in neighbors$T (see moead() for details).

decomp$.nobj: integer value, decomp$.nobj > 1. Number of objectives of the problem.

...

other parameters (included for compatibility with generic call)

Details

This routine calculates the weight vectors for the MOEA/D using the Uniform Design:

References

R. Wang, T. Zhang, B. Guo, "An enhanced MOEA/D using uniform directions and a pre-organization procedure". Proc. IEEE Congress on Evolutionary Computation, Cancun, Mexico, 2013, pp. 2390–2397.

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06

Examples

decomp <- list(name = "uniform", N = 50, .nobj = 3)
W <- decomposition_uniform(decomp)