Differential vector-based local search

Differential vector-based local search (DVLS) implementation for the MOEA/D

ls_dvls(
  Xt,
  Yt,
  Vt,
  B,
  W,
  which.x,
  trunc.x,
  problem,
  scaling,
  aggfun,
  constraint,
  ...
)

Arguments

Xt

Matrix of incumbent solutions

Yt

Matrix of objective function values for Xt

Vt

List object containing information about the constraint violations of the incumbent solutions, generated by evaluate_population()

B

Neighborhood matrix, generated by define_neighborhood().

W

matrix of weights (generated by generate_weights()).

which.x

logical vector indicating which subproblems should undergo local search

trunc.x

logical flag indicating whether candidate solutions generated by local search should be truncated to the variable limits of the problem.

problem

list of named problem parameters. See Section Problem Description of the moead() documentation for details.

scaling

list containing the scaling parameters (see moead() for details).

aggfun

List containing the aggregation function parameters. See Section Scalar Aggregation Functions of the moead() documentation for details.

constraint

list containing the parameters defining the constraint handling method. See Section Constraint Handling of the moead() documentation for details.

...

other parameters (included for compatibility with generic call)

Value

List object with fields X (matrix containing the modified points, with points that did not undergo local search indicated as NA) and nfe

(integer value informing how many additional function evaluations were performed).

Details

This routine implements the differential vector-based local search for the MOEADr package. Check the references for details.

This routine is intended to be used internally by variation_localsearch(), and should not be called directly by the user.

References

B. Chen, W. Zeng, Y. Lin, D. Zhang, "A new local search-based multiobjective optimization algorithm", IEEE Trans. Evolutionary Computation 19(1):50-73, 2015.

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06