preset_moead

Generate a preset configuration for moead()].

preset_moead(name = NULL)

Arguments

name

name of the preset to be generated. Use preset_moead() to obtain the list of available options.

Value

List object containing the preset, to be used as an input to moead(); or, if name == NULL (the default), returns a logical flag invisibly.

Details

This function returns a list of configuration presets taken from the literature to be used with the moead() function in package MOEADr.

Use these configurations as a starting point. We strongly recommend that you play around with the particular configurations (see example).

References

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06

Examples

# Generate list of available presets
preset_moead(name = NULL)
#>        name x                                               description
#> 1 original  | Original MOEA/D: Zhang and Li (2007), Sec. V-E, p.721-722
#> 2 original2 | Original MOEA/D, v2: Zhang and Li (2007), Sec. V-F, p.724
#> 3 moead.de  | MOEA/D-DE: Li and Zhang (2009)                           
#> 
#> 
#> Use preset_moead("name") to generate a standard MOEAD composition
#> 

if (FALSE) {
  library(smoof) # < Install package smoof if needed
  ZDT1 <- make_vectorized_smoof(prob.name  = "ZDT1",
                                dimensions = 30)
                                problem   <- list(name       = "ZDT1",
                                                  xmin       = rep(0, 30),
                                                  xmax       = rep(1, 30),
                                                  m          = 2)

  # Get preset configuration for original MOEA/D
  configuration <- preset_moead("original")

  # Modify whatever you fancy:
  stopcrit <- list(list(name = "maxiter", maxiter = 50))
  showpars <- list(show.iters = "dots", showevery  = 10)
  seed     <- 42

  output <- moead(problem  = problem,
                  preset   = configuration,
                  showpars = showpars,
                  stopcrit = stopcrit,
                  seed     = seed)
}