Print progress of MOEA/D

Echoes progress of MOEA/D to the terminal for the MOEADr package

print_progress(iter.times, showpars)

Arguments

iter.times

vector of iteration times of the moead() routine.

showpars

list object containing parameters that control the printed output of moead(). Parameter showpars can have the following key-value pairs:

$show.iters: type of output ("dots", "numbers", or "none"). If not present in showpars, it defaults to "numbers";

$showevery: positive integer that determines how frequently the routine echoes something to the terminal. If not present in showpars, it defaults to 10.

References

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06