Scalarize values for MOEA/D

Perform scalarization for the MOEADr package.

scalarize_values(normYs, W, B, aggfun)

Arguments

normYs

List generated by scale_objectives(), containing two matrices of scaled objective values (normYs$Y and normYs$Yt) and two vectors, containing the current estimates of the ideal (normYs$minP) and nadir (normYs$maxP) points. See scale_objectives() for details.

W

matrix of weights, generated by generate_weights().

B

neighborhood matrix, generated by define_neighborhood().

aggfun

List containing the aggregation function parameters. See Section Scalar Aggregation Functions of the moead() documentation for details.

Value

[ (T+1) x N ] matrix of scalarized performance values. Each column contains the T scalarized performances of the candidate solutions in the neighborhood of a given subproblem, plus the scalarized performance value for the incumbent solution for that subproblem.

Details

This routine calculates the scalarized performance values for the MOEA/D.

The list of available scalarization methods can be generated using get_scalarization_methods()

References

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06