Stop criterion: maximum runtime

Verifies stop criterion "run time limit" for the MOEADr package. For internal use only, not to be called directly by the user.

stop_maxtime(stopcrit, iter.times, ...)

Arguments

stopcrit

list containing the parameters defining the stop handling method. See Section Constraint Handling of the moead() documentation for details.

iter.times

vector containing the times spent by each iteration of the moead() routine, up to the current one.

...

other parameters (included for compatibility with generic call)

Value

boolean value: TRUE if this criterion has been met, FALSE

otherwise.

Details

When this stop criterion is used, one element of the stopcrit parameter (see moead()) must have the following structure:

stopcrit$name = "maxtime"

stopcrit$maxtime, containing a positive integer representing the desired time limit (in seconds).

Warning

This function uses Sys.time() for verifying the total run time, i.e., it counts wall-clock time, not CPU time.

References

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06