summary.moead

S3 method for summarizing moead objects (the output of moead()).

# S3 method for moead
summary(
  object,
  ...,
  useArchive = FALSE,
  viol.threshold = 1e-06,
  ndigits = 3,
  ref.point = NULL,
  ref.front = NULL
)

Arguments

object

list object of class moead (generated by moead())

...

other parameters to be passed down to specific summary functions (currently unused)

useArchive

logical flag to use information from object$Archive. Only used if object$Archive is not NULL.

viol.threshold

threshold of tolerated constraint violation, used to determine feasibility of points in object.

ndigits

number of decimal places to use for the ideal and nadir estimates

ref.point

reference point for calculating the dominated hypervolume (only if package emoa is available). If NULL the estimated nadir point is used instead.

ref.front

Np x Nobj matrix containing a sample of the true Pareto-optimal front, for calculating IGD.

References

F. Campelo, L.S. Batista, C. Aranha (2020): The MOEADr Package: A Component-Based Framework for Multiobjective Evolutionary Algorithms Based on Decomposition. Journal of Statistical Software doi:10.18637/jss.v092.i06

Examples

problem.1 <- list(name = "example_problem",
                  xmin = rep(-1,30),
                  xmax = rep(1,30),
                  m    = 2)
out <- moead(preset    = preset_moead("original2"),
             problem   = problem.1,
             stopcrit  = list(list(name = "maxiter",
                                   maxiter = 100)),
             showpars  = list(show.iters = "dots",
                              showevery  = 10))
#> 
#> MOEA/D running: ..........
summary(out)
#> Warning: reference point not provided:
#> 
#>           using the maximum in each dimension instead.
#> Summary of MOEA/D run
#> #====================================
#> Total function evaluations:  10100
#> Total iterations:  100
#> Population size:  100
#> Feasible points found:  100 (100% of total)
#> Nondominated points found:  58 (58% of total)
#> Estimated ideal point:  66.266 108.596
#> Estimated nadir point:  115.723 179.052
#> Estimated HV:  1755.871
#> Ref point used for HV:  115.7231 179.0521
#> #====================================