Local search Operators

Local search operators for the MOEA/D

variation_localsearch(...)

Arguments

...
arguments to be passed down to the specific ls_xyz() functions. A list of available local search methods can be generated by get_localsearch_methods(). Consult the documentation of the specific functions for details.

Value

Either a matrix Xls containing the modified points (points that did not undergo local search are indicated as NA in this output matrix), or a list object containing the Xls matrix and an integer nfe, informing how many additional function evaluations were performed by the local search operator. The specific output is defined by the ls_xyz() method used.

Details

This routine calls the local search operator for the MOEADr package, as part of the call to perform_variation(). This operator requires its entry in the variation stack (see Section Variation Operators of moead()) to contain the following fields:

name = "localsearch"

type (see get_localsearch_methods() for details)

gamma.ls (optional): probability of application of local search to a given subproblem at any given iteration (numeric between 0 and 1)

tau.ls (optional): period of application of local search to each subproblem (positive integer)

trunc.x (optional): logical flag for truncating the results of the local search operator to the limits defined by problem$xmin, problem$xmax (logical). Defaults to TRUE.

Whenever local search is triggered for a given subproblem, it cancels all other variation operators for that subproblem and is executed directly on the incumbent solution.

This routine is intended to be used internally by perform_variation(), and should not be called directly by the user.