Three-point quadratic approximation local search

Three-point quadratic approximation (TPQA) local search implementation for the MOEA/D

ls_tpqa(Xt, Yt, W, B, Vt, scaling, aggfun, constraint, epsilon = 1e-06,
  which.x, ...)

Arguments

Xt
Matrix of incumbent solutions

Yt
Matrix of objective function values for Xt

W
matrix of weights (generated by generate_weights()).

B
Neighborhood matrix, generated by define_neighborhood().

Vt
List object containing information about the constraint violations of the incumbent solutions, generated by evaluate_population()

scaling
list containing the scaling parameters (see moead() for details).

aggfun
List containing the aggregation function parameters. See Section Scalar Aggregation Functions of the moead() documentation for details.

constraint
list containing the parameters defining the constraint handling method. See Section Constraint Handling of the moead() documentation for details.

epsilon
threshold for using the quadratic approximation value

which.x
logical vector indicating which subproblems should undergo local search

...
other parameters (included for compatibility with generic call)

Value

Matrix X' containing the modified population

Details

This routine implements the 3-point quadratic approximation local search for the MOEADr package. Check the references for details.

This routine is intended to be used internally by variation_localsearch(), and should not be called directly by the user.

References

Y. Tan, Y. Jiao, H. Li, X. Wang, "A modification to MOEA/D-DE for multiobjective optimization problems with complicated Pareto sets", Information Sciences 213(1):14-38, 2012.

Y.-C. Jiao, C. Dang, Y. Leung, Y. Hao, "A modification to the new version of the prices algorithm for continuous global optimization problems", J. Global Optimization 36(4):609-626, 2006.